fix(core): Switch back to ng-model and $watch to preview refresh purpose
The preview is now force-refreshed everytime the binded variable changes, particularly usefull when two markdown are binded on the same.
Passing the link shortcut into the compile post function for future optimization regarding the elastic directive.

BREAKING CHANGE: Instead of the `content` attribute to bind the value, you should now use the `ng-model`.

Closes #3.